Merge zi_developmentmerge12

+++ org.adempiere.base/src/org/compiere/model/MJournal.java
@@ -33,6 +33,7 @@ import org.compiere.util.DisplayType;
 import org.compiere.util.Env;
 import org.compiere.util.Msg;
 import org.compiere.util.TimeUtil;
+import org.compiere.util.Util;
 
 /**
  *  GL Journal Model
@@ -54,9 +55,21 @@ import org.compiere.util.TimeUtil;
 public class MJournal extends X_GL_Journal implements DocAction
 {
 	/**
-	 * 
+	 * generated serial id
 	 */
 	private static final long serialVersionUID = 4661098755828765138L;
+
+    /**
+     * UUID based Constructor
+     * @param ctx  Context
+     * @param GL_Journal_UU  UUID key
+     * @param trxName Transaction
+     */
+    public MJournal(Properties ctx, String GL_Journal_UU, String trxName) {
+        super(ctx, GL_Journal_UU, trxName);
+		if (Util.isEmpty(GL_Journal_UU))
+			setInitialDefaults();
+    }
 
 	/**
 	 * 	Standard Constructor
@@ -68,20 +81,25 @@ public class MJournal extends X_GL_Journal implements DocAction
 	{
 		super (ctx, GL_Journal_ID, trxName);
 		if (GL_Journal_ID == 0)
-		{
-			setCurrencyRate (Env.ONE);
-			setDateDoc (new Timestamp(System.currentTimeMillis()));
-			setDocAction (DOCACTION_Complete);
-			setDocStatus (DOCSTATUS_Drafted);
-			setPostingType (POSTINGTYPE_Actual);
-			setTotalCr (Env.ZERO);
-			setTotalDr (Env.ZERO);
-			setIsApproved (false);
-			setIsPrinted (false);
-			setPosted (false);
-			setProcessed(false);
-		}
+			setInitialDefaults();
 	}	//	MJournal
+
+	/**
+	 * Set the initial defaults for a new record
+	 */
+	private void setInitialDefaults() {
+		setCurrencyRate (Env.ONE);
+		setDateDoc (new Timestamp(System.currentTimeMillis()));
+		setDocAction (DOCACTION_Complete);
+		setDocStatus (DOCSTATUS_Drafted);
+		setPostingType (POSTINGTYPE_Actual);
+		setTotalCr (Env.ZERO);
+		setTotalDr (Env.ZERO);
+		setIsApproved (false);
+		setIsPrinted (false);
+		setPosted (false);
+		setProcessed(false);
+	}
 
 	/**
 	 * 	Load Constructor
@@ -114,8 +132,8 @@ public class MJournal extends X_GL_Journal implements DocAction
 	
 	/**
 	 * 	Copy Constructor.
-	 * 	Dos not copy: Dates/Period
-	 *	@param original original
+	 * 	Does not copy: Dates/Period.
+	 *	@param original original journal document
 	 */
 	public MJournal (MJournal original)
 	{
@@ -148,6 +166,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	@param AD_Client_ID client
 	 * 	@param AD_Org_ID org
 	 */
+	@Override
 	public void setClientOrg (int AD_Client_ID, int AD_Org_ID)
 	{
 		super.setClientOrg(AD_Client_ID, AD_Org_ID);
@@ -155,7 +174,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 
 	/**
 	 * 	Set Accounting Date.
-	 * 	Set also Period if not set earlier
+	 * 	Set also Period if not set earlier.
 	 *	@param DateAcct date
 	 */
 	public void setDateAcct (Timestamp DateAcct)
@@ -172,7 +191,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 
 	/**
 	 * 	Set Currency Info
-	 *	@param C_Currency_ID currenct
+	 *	@param C_Currency_ID currency
 	 *	@param C_ConversionType_ID type
 	 *	@param CurrencyRate rate
 	 */
@@ -202,14 +221,13 @@ public class MJournal extends X_GL_Journal implements DocAction
 		}
 	}
 	
-	/**************************************************************************
+	/**
 	 * 	Get Journal Lines
-	 * 	@param requery requery (not used)
+	 * 	@param requery ignore
 	 *	@return Array of lines
 	 */
 	public MJournalLine[] getLines (boolean requery)
 	{
-		//FR: [ 2214883 ] Remove SQL code and Replace for Query - red1
 		final String whereClause = "GL_Journal_ID=?";
 		List <MJournalLine> list = new Query(getCtx(), I_GL_JournalLine.Table_Name, whereClause, get_TrxName())
 			.setParameters(getGL_Journal_ID())
@@ -269,6 +287,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Propagate to Lines/Taxes
 	 *	@param processed processed
 	 */
+	@Override
 	public void setProcessed (boolean processed)
 	{
 		super.setProcessed (processed);
@@ -281,12 +300,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 		if (log.isLoggable(Level.FINE)) log.fine(processed + " - Lines=" + noLine);
 	}	//	setProcessed
 
-	
-	/**************************************************************************
-	 * 	Before Save
-	 *	@param newRecord new
-	 *	@return true
-	 */
+	@Override
 	protected boolean beforeSave (boolean newRecord)
 	{
 		if (getGL_JournalBatch_ID() > 0) {
@@ -296,7 +310,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 				return false;
 			}
 		}
-		//	Imported Journals may not have date
+		// Set DateDoc to DateAcct or today date
 		if (getDateDoc() == null)
 		{
 			if (getDateAcct() == null)
@@ -312,7 +326,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 		}
 		else if (!isProcessed())
 		{
-			//validate period
+			// Validate period for DateAcct
 			int C_Period_ID = MPeriod.getC_Period_ID(getCtx(), getDateAcct(), getAD_Org_ID());
 			if (C_Period_ID == 0)
 			{
@@ -328,17 +342,20 @@ public class MJournal extends X_GL_Journal implements DocAction
 			}
 		}
 
+		// Set GL_Category_ID from document type
 		if (getGL_Category_ID() == 0 && getC_DocType_ID() > 0)
 			setGL_Category_ID(MDocType.get(getCtx(), getC_DocType_ID()).getGL_Category_ID());
+		// Set C_AcctSchema_ID to primary accounting schema
 		if (getC_AcctSchema_ID() == 0)
 			setC_AcctSchema_ID(MClientInfo.get(getCtx(), getAD_Client_ID()).getC_AcctSchema1_ID());
+		// Set default currency conversion type
 		if (getC_ConversionType_ID() == 0)
 			setC_ConversionType_ID(MConversionType.getDefault(getAD_Client_ID()));
 
 		// IDEMPIERE-63
-		// for documents that can be reactivated we cannot allow changing 
-		// C_DocTypeTarget_ID or C_DocType_ID if they were already processed and isOverwriteSeqOnComplete
-		// neither change the Date if isOverwriteDateOnComplete
+		// If document have been processed, we can't change 
+		// C_DocTypeTarget_ID or C_DocType_ID if DocType.IsOverwriteSeqOnComplete=Y.
+		// Also, can't change DateDoc if DocType.IsOverwriteDateOnComplete=Y.
 		BigDecimal previousProcessedOn = (BigDecimal) get_ValueOld(COLUMNNAME_ProcessedOn);
 		if (! newRecord && previousProcessedOn != null && previousProcessedOn.signum() > 0) {
 			int previousDocTypeID = (Integer) get_ValueOld(COLUMNNAME_C_DocType_ID);
@@ -357,7 +374,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 			}
 		}
 		
-		// Update DateAcct on lines - teo_sarca BF [ 1775358 ]
+		// Propagate DateAcct change to lines
 		if (is_ValueChanged(COLUMNNAME_DateAcct)) {
 			int no = DB.executeUpdate(
 					"UPDATE GL_JournalLine SET "+MJournalLine.COLUMNNAME_DateAcct+"=? WHERE GL_Journal_ID=?",
@@ -367,15 +384,8 @@ public class MJournal extends X_GL_Journal implements DocAction
 		}
 		return true;
 	}	//	beforeSave
-	
-	
-	/**
-	 * 	After Save.
-	 * 	Update Batch Total
-	 *	@param newRecord true if new record
-	 *	@param success true if success
-	 *	@return success
-	 */
+		
+	@Override
 	protected boolean afterSave (boolean newRecord, boolean success)
 	{
 		if (!success)
@@ -383,11 +393,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 		return updateBatch();
 	}	//	afterSave
 	
-	/**
-	 * 	After Delete
-	 *	@param success true if deleted
-	 *	@return true if success
-	 */
+	@Override
 	protected boolean afterDelete (boolean success)
 	{
 		if (!success)
@@ -396,7 +402,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	}	//	afterDelete
 	
 	/**
-	 * 	Update Batch total
+	 * 	Update Total DR and Total CR of journal batch
 	 *	@return true if ok
 	 */
 	protected boolean updateBatch()
@@ -413,13 +419,13 @@ public class MJournal extends X_GL_Journal implements DocAction
 		}
 		return true;
 	}	//	updateBatch
-	
-	
-	/**************************************************************************
+		
+	/**
 	 * 	Process document
 	 *	@param processAction document action
 	 *	@return true if performed
 	 */
+	@Override
 	public boolean processIt (String processAction)
 	{
 		m_processMsg = null;
@@ -436,6 +442,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Unlock Document.
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean unlockIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -447,6 +454,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Invalidate Document
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean invalidateIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -457,6 +465,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 *	Prepare Document
 	 * 	@return new status (In Progress or Invalid) 
 	 */
+	@Override
 	public String prepareIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -572,6 +581,10 @@ public class MJournal extends X_GL_Journal implements DocAction
 		return DocAction.STATUS_InProgress;
 	}	//	prepareIt
 	
+	/**
+	 * @param dateAcct
+	 * @return null or error message
+	 */
 	private String validatePeriod(Timestamp dateAcct) {
 		// Get Period
 		MDocType dt = MDocType.get(getCtx(), getC_DocType_ID());
@@ -605,6 +618,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Approve Document
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean  approveIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -616,6 +630,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Reject Approval
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean rejectIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -627,6 +642,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Complete Document
 	 * 	@return new status (Complete, In Progress, Invalid, Waiting ..)
 	 */
+	@Override
 	public String completeIt()
 	{
 		//	Re-Check
@@ -690,6 +706,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Void Document.
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean voidIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -735,6 +752,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Cancel not delivered Qunatities
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean closeIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -762,10 +780,11 @@ public class MJournal extends X_GL_Journal implements DocAction
 	}	//	closeIt
 	
 	/**
-	 * 	Reverse Correction (in same batch).
-	 * 	As if nothing happened - same date
+	 * 	Reverse Correction (in same batch if using batch).
+	 *  Flip Dr/Cr - Use date of this document.
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean reverseCorrectIt()
 	{
 		// Before reverseCorrect
@@ -788,8 +807,8 @@ public class MJournal extends X_GL_Journal implements DocAction
 
 	/**
 	 * 	Reverse Correction.
-	 * 	As if nothing happened - same date
-	 * 	@param GL_JournalBatch_ID reversal batch
+	 *  Flip Dr/Cr - Use date of this document.
+	 * 	@param GL_JournalBatch_ID optional reversal batch
 	 * 	@return reversed Journal or null
 	 */
 	public MJournal reverseCorrectIt (int GL_JournalBatch_ID)
@@ -837,10 +856,11 @@ public class MJournal extends X_GL_Journal implements DocAction
 	}	//	reverseCorrectionIt
 	
 	/**
-	 * 	Reverse Accrual (sane batch).
-	 * 	Flip Dr/Cr - Use Today's date
+	 * 	Reverse Accrual (in same batch if using batch).
+	 * 	Flip Dr/Cr - Use Today's date.
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean reverseAccrualIt()
 	{
 		// Before reverseAccrual
@@ -863,8 +883,8 @@ public class MJournal extends X_GL_Journal implements DocAction
 	
 	/**
 	 * 	Reverse Accrual.
-	 * 	Flip Dr/Cr - Use Today's date
-	 * 	@param GL_JournalBatch_ID reversal batch
+	 * 	Flip Dr/Cr - Use Today's date.
+	 * 	@param GL_JournalBatch_ID optional reversal batch
 	 * 	@return reversed journal or null 
 	 */
 	public MJournal reverseAccrualIt (int GL_JournalBatch_ID)
@@ -911,9 +931,10 @@ public class MJournal extends X_GL_Journal implements DocAction
 	}	//	reverseAccrualIt
 	
 	/** 
-	 * 	Re-activate
+	 * 	Re-activate document and delete Fact_Acct entries.
 	 * 	@return true if success 
 	 */
+	@Override
 	public boolean reActivateIt()
 	{
 		if (log.isLoggable(Level.INFO)) log.info(toString());
@@ -921,9 +942,15 @@ public class MJournal extends X_GL_Journal implements DocAction
 		m_processMsg = ModelValidationEngine.get().fireDocValidate(this,ModelValidator.TIMING_BEFORE_REACTIVATE);
 		if (m_processMsg != null)
 			return false;	
-		
+
 		// teo_sarca - FR [ 1776045 ] Add ReActivate action to GL Journal
 		MPeriod.testPeriodOpen(getCtx(), getDateAcct(), getC_DocType_ID(), getAD_Org_ID());
+
+		if (!DocumentEngine.canReactivateThisDocType(getC_DocType_ID())) {
+			m_processMsg = Msg.getMsg(getCtx(), "DocTypeCannotBeReactivated", new Object[] {MDocType.get(getC_DocType_ID()).getNameTrl()});
+			return false;
+		}
+
 		MFactAcct.deleteEx(MJournal.Table_ID, get_ID(), get_TrxName());
 		setPosted(false);
 		setProcessed(false);
@@ -936,12 +963,12 @@ public class MJournal extends X_GL_Journal implements DocAction
 		
 		return true;
 	}	//	reActivateIt
-	
-	
-	/*************************************************************************
+		
+	/**
 	 * 	Get Summary
 	 *	@return Summary of Document
 	 */
+	@Override
 	public String getSummary()
 	{
 		StringBuilder sb = new StringBuilder();
@@ -962,6 +989,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	String Representation
 	 *	@return info
 	 */
+	@Override
 	public String toString ()
 	{
 		StringBuilder sb = new StringBuilder ("MJournal[");
@@ -976,6 +1004,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Get Document Info
 	 *	@return document info (untranslated)
 	 */
+	@Override
 	public String getDocumentInfo()
 	{
 		MDocType dt = MDocType.get(getCtx(), getC_DocType_ID());
@@ -987,6 +1016,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Create PDF
 	 *	@return File or null
 	 */
+	@Override
 	public File createPDF ()
 	{
 		try
@@ -1005,18 +1035,18 @@ public class MJournal extends X_GL_Journal implements DocAction
 	/**
 	 * 	Create PDF file
 	 *	@param file output file
-	 *	@return file if success
+	 *	@return not implemented, always return null
 	 */
 	public File createPDF (File file)
 	{
 		return null;
 	}	//	createPDF
-
 	
 	/**
 	 * 	Get Process Message
 	 *	@return clear text error message
 	 */
+	@Override
 	public String getProcessMsg()
 	{
 		return m_processMsg;
@@ -1026,6 +1056,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Get Document Owner (Responsible)
 	 *	@return AD_User_ID (Created)
 	 */
+	@Override
 	public int getDoc_User_ID()
 	{
 		return getCreatedBy();
@@ -1035,6 +1066,7 @@ public class MJournal extends X_GL_Journal implements DocAction
 	 * 	Get Document Approval Amount
 	 *	@return DR amount
 	 */
+	@Override
 	public BigDecimal getApprovalAmt()
 	{
 		return getTotalDr();
@@ -1053,8 +1085,8 @@ public class MJournal extends X_GL_Journal implements DocAction
 	}	//	isComplete
 
 	/**
-	 * 	Get Document Status
-	 *	@return Document Status Clear Text
+	 * 	Get Document Status Name
+	 *	@return Document Status Name
 	 */
 	public String getDocStatusName()
 	{
